refactor (gravsearch): move gravsearch Sparql transformer class into separate file
