+++ webapi/src/main/scala/org/knora/webapi/util/search/gravsearch/NonTriplestoreSpecificConstructToSelectTransformerCountQuery.scala
@@ -72,5 +72,4 @@ class NonTriplestoreSpecificConstructToSelectTransformerCountQuery(typeInspectio
         // count queries do not consider offsets since there is only one result row
         0
     }
-
 }
